requirements.txt
streamlit==1.28.0
numpy==1.24.3
pillow==10.1.0
matplotlib==3.7.2
scikit-learn==1.3.0